Attempt to improve documentation.
I don't think my `--help` doc for `--fundamental-packages` is much
better than it was before, but "already being installed" is the only
property of "fundamental" packages I could come up with. Maybe John
could write something more clear?

It would also be nice if `--help` included the "usage", but I didn't
see any option for that in the docs for the `options` package.