Add option to consider all globally installed packages "fundamental".
And rename "fundamental" to "global", since that seems to be what
"fundamental" means here: the "fundamental" packages must be globally
installed, so just say that.

With this I can further simplify my usage. Now either of these work
for me:

    cabal-graphdeps/cabal-graphdeps --cabal halvm-cabal --ghc-pkg halvm-ghc-pkg --add-sources .,hsntp,ssh-hans,ssh-hans-chatter --cabal-config halvm-cabal-sandbox/cabal.config prattle --global-packages base,HALVMCore,XenDevice,binary,containers,time

and

    cabal-graphdeps --cabal halvm-cabal --ghc-pkg halvm-ghc-pkg --add-sources .,hsntp,ssh-hans,ssh-hans-chatter --cabal-config halvm-cabal-sandbox/cabal.config prattle --use-all-global-package

They produce different output, because the first version doesn't list
the entire global package db, just enough packages to make dependency
resolution work. A better version of `--use-all-global-packages` might
simply pin the versions of those packages to the globally installed
versions, but still make them appear in the output as deps.